requirements.txt
amqp==1.4.9
anyjson==0.3.3
bcrypt==3.1.0
billiard==3.3.0.23
blinker==1.4
celery==3.1.23
cffi==1.7.0
click==6.6
coverage==4.2
fake-factory==0.5.9
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Redis==0.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
MarkupSafe==0.23
misaka==2.0.0
MySQL-python==1.2.5
Pillow==3.3.0
pycparser==2.14
python-dateutil==2.5.3
pytz==2016.4
qiniu==7.0.7
redis==2.10.5
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.13
Werkzeug==0.11.10
WTForms==2.1
